EXAMPLE: overridable vue component templates
This commit shows how we can vue components compatible with the joomla override system.

There are some drawbacks:
We cant use the vue runtime-only build anymore and have to use the standalone build, which does include the template compiler. The runtime-only build is roughly 30% lighter-weight than the standalone build, weighing only 18.01kb min+gzip.

It is harder to develop and maintain, when components and templates are separated across multiple files.